Moved Article type from article-api to common
