+++ article-api/src/main/scala/no/ndla/articleapi/model/api/ArticleDomainDump.scala
@@ -8,7 +8,7 @@
 package no.ndla.articleapi.model.api
 
 import org.scalatra.swagger.annotations.{ApiModel, ApiModelProperty}
-import no.ndla.articleapi.model.domain
+import no.ndla.common.model.domain.article.Article
 
 import scala.annotation.meta.field
 
@@ -17,5 +17,5 @@ case class ArticleDomainDump(
     @(ApiModelProperty @field)(description = "The total number of articles in the database") totalCount: Long,
     @(ApiModelProperty @field)(description = "For which page results are shown from") page: Int,
     @(ApiModelProperty @field)(description = "The number of results per page") pageSize: Int,
-    @(ApiModelProperty @field)(description = "The search results") results: Seq[domain.Article]
+    @(ApiModelProperty @field)(description = "The search results") results: Seq[Article]
 )
