+++ article-api/src/main/scala/no/ndla/articleapi/model/domain/Content.scala
@@ -9,52 +9,13 @@
 package no.ndla.articleapi.model.domain
 
 import no.ndla.articleapi.Props
-import no.ndla.common.model.domain.{
-  Introduction,
-  Description,
-  Availability,
-  RelatedContent,
-  RequiredLibrary,
-  Tag,
-  Title,
-  VisualElement
-}
-import no.ndla.common.model.domain.article.Copyright
+import no.ndla.common.model.domain.Availability
+import no.ndla.common.model.domain.article.Article
 import org.json4s.{DefaultFormats, FieldSerializer, Formats}
 import org.json4s.FieldSerializer._
 import org.json4s.ext.{EnumNameSerializer, JavaTimeSerializers}
 import org.json4s.native.Serialization._
 import scalikejdbc._
-
-import java.time.LocalDateTime
-
-sealed trait Content {
-  def id: Option[Long]
-}
-
-case class Article(
-    id: Option[Long],
-    revision: Option[Int],
-    title: Seq[Title],
-    content: Seq[ArticleContent],
-    copyright: Copyright,
-    tags: Seq[Tag],
-    requiredLibraries: Seq[RequiredLibrary],
-    visualElement: Seq[VisualElement],
-    introduction: Seq[Introduction],
-    metaDescription: Seq[Description],
-    metaImage: Seq[ArticleMetaImage],
-    created: LocalDateTime,
-    updated: LocalDateTime,
-    updatedBy: String,
-    published: LocalDateTime,
-    articleType: String,
-    grepCodes: Seq[String],
-    conceptIds: Seq[Long],
-    availability: Availability.Value = Availability.everyone,
-    relatedContent: Seq[RelatedContent],
-    revisionDate: Option[LocalDateTime]
-) extends Content
 
 trait DBArticle {
   this: Props =>
