+++ article-api/src/main/scala/no/ndla/articleapi/service/search/SearchConverterService.scala
@@ -10,9 +10,9 @@ package no.ndla.articleapi.service.search
 
 import com.typesafe.scalalogging.StrictLogging
 import no.ndla.articleapi.model.api.{ArticleSummaryV2, SearchResultV2}
-import no.ndla.articleapi.model.domain._
 import no.ndla.articleapi.model.search._
 import no.ndla.articleapi.service.ConverterService
+import no.ndla.common.model.domain.article.Article
 import no.ndla.search.SearchLanguage.languageAnalyzers
 import no.ndla.search.model.{LanguageValue, SearchableLanguageList, SearchableLanguageValues}
 import org.jsoup.Jsoup
