+++ article-api/src/main/scala/no/ndla/articleapi/validation/ContentValidator.scala
@@ -10,10 +10,19 @@ package no.ndla.articleapi.validation
 
 import no.ndla.articleapi.Props
 import no.ndla.articleapi.integration.DraftApiClient
-import no.ndla.articleapi.model.domain._
 import no.ndla.common.errors.{ValidationException, ValidationMessage}
-import no.ndla.common.model.domain.{ArticleType, Author, Description, Introduction, RequiredLibrary, Tag, VisualElement}
-import no.ndla.common.model.domain.article.Copyright
+import no.ndla.common.model.domain.{
+  ArticleContent,
+  ArticleMetaImage,
+  ArticleType,
+  Author,
+  Description,
+  Introduction,
+  RequiredLibrary,
+  Tag,
+  VisualElement
+}
+import no.ndla.common.model.domain.article.{Article, Copyright}
 import no.ndla.language.model.{Iso639, LanguageField}
 import no.ndla.mapping.License.getLicense
 import no.ndla.validation.HtmlTagRules.stringToJsoupDocument
