+++ article-api/src/test/scala/no/ndla/articleapi/controller/InternControllerTest.scala
@@ -8,8 +8,8 @@
 
 package no.ndla.articleapi.controller
 
-import no.ndla.articleapi.model.domain._
 import no.ndla.articleapi.{TestEnvironment, UnitSuite}
+import no.ndla.common.model.domain.article.Article
 import no.ndla.common.model.domain.{Author, Availability}
 import org.json4s.{DefaultFormats, Formats}
 import org.json4s.ext.{EnumNameSerializer, JavaTimeSerializers}
