+++ article-api/src/test/scala/no/ndla/articleapi/model/search/SearchableArticleSerializerTest.scala
@@ -8,8 +8,7 @@
 package no.ndla.articleapi.model.search
 
 import no.ndla.articleapi._
-import no.ndla.articleapi.model.domain.ArticleMetaImage
-import no.ndla.common.model.domain.Availability
+import no.ndla.common.model.domain.{ArticleMetaImage, Availability}
 import no.ndla.search.model.{LanguageValue, SearchableLanguageFormats, SearchableLanguageList, SearchableLanguageValues}
 import org.json4s.Formats
 import org.json4s.native.Serialization.{read, writePretty}
