+++ article-api/src/test/scala/no/ndla/articleapi/service/WriteServiceTest.scala
@@ -8,8 +8,8 @@
 
 package no.ndla.articleapi.service
 
-import no.ndla.articleapi.model.domain._
 import no.ndla.articleapi.{TestEnvironment, UnitSuite}
+import no.ndla.common.model.domain.article.Article
 import org.mockito.invocation.InvocationOnMock
 import org.mockito.{ArgumentCaptor, Mockito}
 import scalikejdbc.DBSession
