+++ article-api/src/test/scala/no/ndla/articleapi/validation/ContentValidatorTest.scala
@@ -8,10 +8,18 @@
 
 package no.ndla.articleapi.validation
 
-import no.ndla.articleapi.model.domain._
 import no.ndla.articleapi.{TestEnvironment, UnitSuite}
 import no.ndla.common.errors.{ValidationException, ValidationMessage}
-import no.ndla.common.model.domain.{Author, Description, Introduction, RequiredLibrary, Tag, Title}
+import no.ndla.common.model.domain.{
+  ArticleContent,
+  ArticleMetaImage,
+  Author,
+  Description,
+  Introduction,
+  RequiredLibrary,
+  Tag,
+  Title
+}
 import no.ndla.common.model.domain.article.Copyright
 import no.ndla.mapping.License.{CC_BY_SA, NA}
